requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
faiss-cpu
requests
python-dotenv
beautifulsoup4
openai
